2dcontext/drawing-text-to-the-canvas/2d.text.measure.offscreencanvas.width.empty.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.text.measure.width.empty</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("/fonts/CanvasTest.ttf");
}
</style>
<body class="show_output">

<h1>2d.text.measure.offscreencanvas.width.empty</h1>
<p class="desc">Test if the empty string has zero width for OffscreenCanvas</p>

<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>

<ul id="d"></ul>
<script>
var t = async_test("The empty string has zero width for OffscreenCanvas");
test(function () {
  deferTest();
  var f = new FontFace("CanvasTest", "/fonts/CanvasTest.ttf");
  document.fonts.add(f);
  document.fonts.ready.then(() => {
    step_timeout(t.step_func_done(function () {
      var offscreen = new OffscreenCanvas(100, 50);
      var ctx = offscreen.getContext('2d');
      ctx.font = '50px CanvasTest';
      _assertSame(ctx.measureText("").width, 0, "ctx.measureText(\"\").width", "0");
    }), 0);
  });
});
</script>
</body>>
